services/image-generation-service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3
pydantic-settings==2.7.0
celery==5.4.0
redis==5.2.1
httpx==0.28.1
PyJWT==2.10.1
python-dotenv==1.0.1
structlog==24.4.0
tenacity==9.0.0
google-cloud-aiplatform==1.78.0
google-cloud-storage==2.19.0
Pillow==11.1.0
pytest==8.3.4
pytest-asyncio==0.24.0
